Harden against crashes caused by account deletes
The basic failure mode is to delete an account while it is being
synced or some other background activity.  The fixes in this CL will
harden against this, but it will not be a perfect solution; That
will require shutting down all sync, which is difficult in the current
too-many-threads implementation.

Primary bug# 2112564