Watch for null results when updating internaldate.
Fixes bug # 2148971

Note, this condition happens when we APPEND a message to the server
but then we turn around and query for it, and the server reports that
it's not there:

 >>> 7 APPEND "Sent" {490}
 <<< #null# [Ready, for, argument]
 <<< #null# [194, EXISTS]
 <<< #7# [OK, [APPENDUID, 1233339552, 1191], APPEND, completed]
 >>> 8 UID SEARCH (HEADER MESSAGE-ID \
       <pwyc0i5yviwvu5nwio2lwdo5.1254073239445@email.android.com>)
 <<< #null# [SEARCH]
 <<< #8# [OK, UID, SEARCH, completed]
 >>> 9 UID SEARCH UID NULL
 <<< #9# [BAD, UID, SEARCH, error;, bogus, or, incorrect, syntax, \
       of, search, critera]

I don't know why the Yahoo! server did this, but now we'll at least not
crash if they do it again.