MailboxList: Show the rotating icon when there's no mailbox
If the cursor is empty, hide the list and let the framework show
the rotating icon.  (Don't set an empty cursor, which makes the list just
blank.)

We do this for only MailboxList because we know we'll get mailboxes
soon enough.  We can't do this on MessageList because there may be really
no messages even if we wait.

Bug 2927973

Change-Id: Ic0346d2d90d03489072eb95a123281f60cda44dd