Simplify drag-n-drop scrolling logic
Remove the concept of scroll speeds. According to Mark, multiple scroll speeds
is a "nice to have" feature and not a "must have". Besides, this did not work
(there was only ever one scroll speed), and, when fixed, the top scroll speed
was way too fast to be usable.

bug 3335062

Change-Id: I160bec2ca5de0b26da81ed31dfb96fc86e543988