update mailbox adapter callback
the callback method name was not appropriate in all situations. This has been
renamed to just "onBind()" and will still be invoked whenever data is bound to
the adapter. also, create an empty callback that can be used if the adapter
doesn't need to be informed when a bind occurs.

Change-Id: Ifaa4aea3a59734557651270330ca9439e3ea4ab4