Always display subfolder icon in the same view
We were handling the subfolder icon strangely. Sometimes it was shown in the
folder_icon view and sometimes in the folder_expanded_icon view. Now the
folder_icon view is used exclusively for system folders (such as inbox, etc...)
and folder_expanded_icon is always used for the expanded/collapsed icon.

Change-Id: Ifff9a44cc792497a469c2408337dafd431f6dfc3