Create new mailbox fragment for nested folder navigation
Create a new instance of the mailbox list fragment when navigating through
nested folders. (drill-in / going back to the root mailboxes.)

Also the fragment manager now has two public methdos for navigation that
are called by the activity.  They will be the common interface for
the tablet UI manager (i.e. MessageListXLFragmentManager) and the
phone UI manager.
- openAccount(accountId)
  Open the default view for the account.  Used when switching accounts.
- open(long accountId, long mailboxId, long messageId)
  Opne a particular view.  Used when, for example, opening a message from the
  widget.

Known issue:
- This breaks drag&drop through nested folder navigation.
  This is because a new mailbox list fragment created during D&D doesn't
  get the "drag started" event.

Change-Id: I69c14b71b4f681f8ab57f3ddd2cff9744a832076