Make callbacks from shortcut picker
We need to move some of the calls from shortcut picker to the activity (such
as creating the actual shortcut and terminating the activity). This should
let us to reuse the picker fragment from the widget code.

Maybe other callbacks are needed, but, this is a first pass to get something
that "should" work.

Change-Id: I6467979951ad73b5817ba543683f62f831ac73a4